Pass row number on queries (MLDB-XXXX)

+++ server/bound_queries.h
@@ -103,7 +103,7 @@ struct BoundSelectQuery {
                      const SqlExpression & where,
                      const OrderByExpression & orderBy,
                      std::vector<std::shared_ptr<SqlExpression> > calc,
-                     int numBuckets = -1);
+                     int64_t numBuckets = -1);
 
     bool execute(RowProcessorEx processor,
                  ssize_t offset,
@@ -111,7 +111,8 @@ struct BoundSelectQuery {
                  std::function<bool (const Json::Value &)> onProgress);
 
     bool execute(std::function<bool (NamedRowValue & output,
-                                     std::vector<ExpressionValue> & calcd, int rowNum)> processor,
+                                     std::vector<ExpressionValue> & calcd,
+                                     int64_t rowNum)> processor,
                  bool processInParallel,
                  ssize_t offset,
                  ssize_t limit,
@@ -125,7 +126,7 @@ struct BoundSelectQuery {
     bool executeExpr(std::function<bool (RowPath & rowName,
                                          ExpressionValue & val,
                                          std::vector<ExpressionValue> & calcd,
-                                         int rowNum)> processor,
+                                         int64_t rowNum)> processor,
                      bool processInParallel,
                      ssize_t offset,
                      ssize_t limit,
@@ -153,9 +154,10 @@ struct BoundGroupByQuery {
                      const SqlExpression & rowName,
                      const OrderByExpression & orderBy);
 
-    std::pair<bool, std::shared_ptr<ExpressionValueInfo> > execute(RowProcessor processor,  
-                     ssize_t offset, ssize_t limit,
-                     std::function<bool (const Json::Value &)> onProgress);
+    std::pair<bool, std::shared_ptr<ExpressionValueInfo> >
+    execute(RowProcessor processor,  
+            ssize_t offset, ssize_t limit,
+            std::function<bool (const Json::Value &)> onProgress);
 
     const Dataset & from;
     WhenExpression when;
